+++ include/sdf/Atmosphere.hh
@@ -20,10 +20,14 @@
 #include <ignition/math/Temperature.hh>
 #include "sdf/Element.hh"
 #include "sdf/Types.hh"
+#include "sdf/sdf_config.h"
 #include "sdf/system_util.hh"
 
 namespace sdf
 {
+  // Inline bracket to help doxygen filtering.
+  inline namespace SDF_VERSION_NAMESPACE {
+  //
   /// \enum AtmosphereType
   /// \brief The set of atmosphere model types.
   enum class AtmosphereType
@@ -106,5 +110,6 @@ namespace sdf
     /// \brief Private data pointer.
     private: AtmospherePrivate *dataPtr = nullptr;
   };
+  }
 }
 #endif