+++ include/sdf/Visual.hh
@@ -26,10 +26,15 @@
 #include "sdf/Plane.hh"
 #include "sdf/Sphere.hh"
 #include "sdf/Types.hh"
+#include "sdf/sdf_config.h"
 #include "sdf/system_util.hh"
 
 namespace sdf
 {
+  // Inline bracket to help doxygen filtering.
+  inline namespace SDF_VERSION_NAMESPACE {
+  //
+
   // Forward declarations.
   class VisualPrivate;
   class Geometry;
@@ -106,5 +111,6 @@ namespace sdf
     /// \brief Private data pointer.
     private: VisualPrivate *dataPtr = nullptr;
   };
+  }
 }
 #endif